App => Design Improvement | v0.9
